% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotMeanSd.R
\name{ggPlotMeanSd}
\alias{ggPlotMeanSd}
\title{ggPlotMeanSd}
\usage{
ggPlotMeanSd(df_xts, l1 = T, l2 = T, l3 = F, mean_col = "black",
  mean_size = 0.8, sd_col = "grey70")
}
\arguments{
\item{sd_col}{}
}
\description{
ggPlotMeanSd
}
