% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.xts.R
\name{mplot.xts}
\alias{mplot.xts}
\title{mplot.xts}
\usage{
\method{mplot}{xts}(..., period = "::", theme = mytheme_right,
  titles = NULL, use.one.x = T, fill.tail = F)
}
\arguments{
\item{fill.tail}{}
}
\description{
mplot.xts
}
