\name{myplot-package}
\alias{myplot-package}
\alias{myplot}
\docType{package}
\title{
\packageTitle{myplot}
}
\description{
\packageDescription{myplot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myplot}
\packageIndices{myplot}
Saving base R graphics like **ggsave**
}
\author{
\packageAuthor{myplot}

Maintainer: \packageMaintainer{myplot}
}

