% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotFunctions}
\alias{plotFunctions}
\title{plotFunctions}
\usage{
plotFunctions(fun, params = NULL, n = 101, x = c(1, 5), size = 0.5,
  end_spaces = 0)
}
\arguments{
\item{fun}{function to plot}

\item{params}{the function params}

\item{n}{dots}

\item{x}{range of x}

\item{size}{line width}

\item{end_spaces}{make some space for label at the end of lines}
}
\value{
ggplot object
}
\description{
plotFunctions
}
\examples{

CRRA = function(x, theta){
  if (theta == 1) {
    return(log(x))
  }
  (x ^ (1 - theta) - 1) / (1 - theta)
}
params = data.frame(theta = c(0, 0.5, 1, 2, 4))
plotFunctions(CRRA, params, x = c(0.1, 5), size = 0.6, end_spaces = 0.05) +
  ggtitle("CRRA") + ylim(-1, 4) + MyPlot::gghline(0)

}
